requirements/base.txt
Django==1.10.6
djangorestframework==3.6.2
django-bootstrap3
django-annoying